refactor: rename AsyncAPISchemaCompletionContributor to AsyncAPISpecificationCompletionContributor
