+++ src/main/resources/META-INF/plugin.xml
@@ -52,8 +52,8 @@
         <psi.referenceContributor language="yaml" implementation="com.asyncapi.plugin.idea.extensions.psi.reference.contributor.yaml.AsyncAPISchemaReferenceContributor"/>
 
         <!-- AsyncAPI schema completion -->
-        <completion.contributor language="JSON" implementationClass="com.asyncapi.plugin.idea.extensions.completion.AsyncAPISchemaCompletionContributor"/>
-        <completion.contributor language="yaml" implementationClass="com.asyncapi.plugin.idea.extensions.completion.AsyncAPISchemaCompletionContributor"/>
+        <completion.contributor language="JSON" implementationClass="com.asyncapi.plugin.idea.extensions.completion.AsyncAPISpecificationCompletionContributor"/>
+        <completion.contributor language="yaml" implementationClass="com.asyncapi.plugin.idea.extensions.completion.AsyncAPISpecificationCompletionContributor"/>
 
         <!-- Preview in built-in/external browser -->
         <webBrowserUrlProvider implementation="com.asyncapi.plugin.idea.extensions.web.AsyncAPIBrowserUrlProvider"/>
