--- src/test/kotlin/com/asyncapi/plugin/idea/completion/v2_1_0/AsyncAPISchemaCompletionContributorJsonTest.kt
@@ -1,11 +1,0 @@
-package com.asyncapi.plugin.idea.completion.v2_1_0
-
-import com.asyncapi.plugin.idea.completion.AbstractAsyncAPISchemaCompletionContributorTest
-
-class AsyncAPISchemaCompletionContributorJsonTest: AbstractAsyncAPISchemaCompletionContributorTest() {
-
-    override fun fileExtension(): String = "json"
-
-    override fun asyncAPIVersion(): String = "2.1.0"
-
-}
